Move ChainApi to Core
