+++ chain/src/main/scala/org/bitcoins/chain/blockchain/ChainHandler.scala
@@ -1,7 +1,6 @@
 package org.bitcoins.chain.blockchain
 
 import org.bitcoins.chain.ChainVerificationLogger
-import org.bitcoins.chain.api.ChainApi
 import org.bitcoins.chain.config.ChainAppConfig
 import org.bitcoins.chain.models._
 import org.bitcoins.chain.pow.Pow
@@ -24,7 +23,7 @@ import scala.concurrent._
 
 /**
   * Chain Handler is meant to be the reference implementation
-  * of [[org.bitcoins.chain.api.ChainApi ChainApi]], this is the entry point in to the
+  * of [[ChainApi ChainApi]], this is the entry point in to the
   * chain project.
   *
   * @param blockHeaderDAO block header DB
