+++ chain/src/main/scala/org/bitcoins/chain/blockchain/sync/ChainSync.scala
@@ -1,10 +1,9 @@
 package org.bitcoins.chain.blockchain.sync
 
 import org.bitcoins.chain.ChainVerificationLogger
-import org.bitcoins.chain.api.ChainApi
 import org.bitcoins.chain.blockchain.ChainHandler
 import org.bitcoins.chain.config.ChainAppConfig
-import org.bitcoins.core.api.chain.db.BlockHeaderDb
+import org.bitcoins.core.api.chain.db.{BlockHeaderDb, ChainApi}
 import org.bitcoins.core.protocol.blockchain.BlockHeader
 import org.bitcoins.crypto.DoubleSha256DigestBE
