+++ core/src/main/scala/org/bitcoins/core/api/chain/db/ChainApi.scala
@@ -1,11 +1,6 @@
-package org.bitcoins.chain.api
+package org.bitcoins.core.api.chain.db
 
 import org.bitcoins.core.api.chain.ChainQueryApi
-import org.bitcoins.core.api.chain.db.{
-  BlockHeaderDb,
-  CompactFilterDb,
-  CompactFilterHeaderDb
-}
 import org.bitcoins.core.gcs.FilterHeader
 import org.bitcoins.core.p2p.CompactFilterMessage
 import org.bitcoins.core.protocol.BlockStamp
