+++ node/src/main/scala/org/bitcoins/node/networking/peer/DataMessageHandler.scala
@@ -1,7 +1,7 @@
 package org.bitcoins.node.networking.peer
 
-import org.bitcoins.chain.api.ChainApi
 import org.bitcoins.chain.config.ChainAppConfig
+import org.bitcoins.core.api.chain.db.ChainApi
 import org.bitcoins.core.gcs.BlockFilter
 import org.bitcoins.core.p2p._
 import org.bitcoins.crypto.DoubleSha256DigestBE
