+++ node/src/main/scala/org/bitcoins/node/networking/peer/PeerMessageReceiver.scala
@@ -1,7 +1,6 @@
 package org.bitcoins.node.networking.peer
 
 import akka.actor.ActorRefFactory
-import org.bitcoins.chain.api.ChainApi
 import org.bitcoins.chain.blockchain.ChainHandler
 import org.bitcoins.chain.config.ChainAppConfig
 import org.bitcoins.chain.models.{
@@ -9,6 +8,7 @@ import org.bitcoins.chain.models.{
   CompactFilterDAO,
   CompactFilterHeaderDAO
 }
+import org.bitcoins.core.api.chain.db.ChainApi
 import org.bitcoins.core.p2p.{NetworkMessage, _}
 import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.node.models.Peer
