+++ testkit/src/main/scala/org/bitcoins/testkit/node/NodeUnitTest.scala
@@ -3,11 +3,11 @@ package org.bitcoins.testkit.node
 import java.net.InetSocketAddress
 
 import akka.actor.{ActorSystem, Cancellable}
-import org.bitcoins.chain.api.ChainApi
 import org.bitcoins.chain.config.ChainAppConfig
 import org.bitcoins.core.api.chain.ChainQueryApi
 import org.bitcoins.core.api.chain.db.{
   BlockHeaderDb,
+  ChainApi,
   CompactFilterDb,
   CompactFilterHeaderDb
 }
